Removed classes target
In order to have the following syntax:
```rust
struct D {}
impl D {
  fn foo(&self) {}
}
fn main() {
   let d = D{};
   d.foo();
   d.hotpatch(PATH);
   d.foo();
}
```
The `foo` method needs to be transmuted into a functor.

This means, interanally, it must be a member of either the struct itself
or a trait. `proc_macro`s can't reach the source struct definition,
and even if there was `#[derive(Patchable)]` the `impl` and `struct` would need
to be in the same `proc_macro` call -- something that goes against Rust's goals.

The best way forward is a trait member... but those aren't in rust yet and have been
posponed for years at this point. It looks like `#[patchable]` won't be able to be
placed anywhere near an `impl` for a long time. Sad.